requirements.txt
FastAPI
pymongo[srv]
dnspython
pandas
openpyxl
python-dotenv
uvicorn